+++ FoodOrderApp/FoodOrderApp/Models/TransportationType.cs
@@ -5,7 +5,7 @@ namespace FoodOrderApp.Models
 {
     public partial class TransportationType
     {
-        public long TransportationId { get; set; }
+        public int TransportationId { get; set; }
         public string TransportationName { get; set; }
         public string DriversLicense { get; set; }
         public bool IsVerified { get; set; }
